+++ event.go
@@ -55,6 +55,7 @@ const (
 	EventA2AAgentRegistered            EventType = "a2a_agent_registered"
 	EventA2AAgentCardFetched           EventType = "a2a_agent_card_fetched"
 	EventA2AAgentHeartbeat             EventType = "a2a_agent_heartbeat"
+	EventA2AAgentEvicted               EventType = "a2a_agent_evicted"
 	EventA2ATaskSent                   EventType = "a2a_task_sent"
 	EventA2AMessageReceived            EventType = "a2a_message_received"
 	EventA2AArtifactUpdated            EventType = "a2a_artifact_updated"